[Payments] - Move CheckInfo & CreditCardInfo into payments folder
To avoid unneeded indirection
