+++ app/components/payments/CreditCardInfo.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Label, Input } from 'reactstrap';
-import styles from './TradingPost.css';
+import styles from '../TradingPost.css';
 
 const CreditCardInfo = ({ title }) => (
     <div className={styles.leftsidehdr}>
-        <h3 >{title}</h3>
+        <h3>Credit Card Payment Info</h3>
         <Input type="textarea" name="text" id="creditCardText" />
     </div>
 );
