--- app/containers/CheckPaymentContainer.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types'
-import CheckInfo from '../components/CheckInfo';
-
-const CheckPaymentContainer = () => (
-  <CheckInfo title="Check Payment"> 
-  </CheckInfo>
-);
-
-export default CheckPaymentContainer
