--- app/containers/CreditCardContainer.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types'
-import CreditCardInfo from '../components/CreditCardInfo';
-
-const CreditCardContainer = () => (
-  <CreditCardInfo title="Credit Card Payment Info"> 
-  </CreditCardInfo>
-);
-
-export default CreditCardContainer
