+++ app/containers/PaymentPage.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Container, Row, Col } from 'reactstrap';
 import { connect } from 'react-redux';
-import CheckPaymentContainer from './CheckPaymentContainer';
-import CreditCardContainer from './CreditCardContainer';
 import CartTotalContainer from './CartTotalContainer';
 import styles from '../components/TradingPost.css';
 import CashPaymentContainer from './CashPaymentContainer';
 import CouncilSearch from '../components/payments/CouncilSearch';
+import CheckInfo from '../components/payments/CheckInfo';
+import CreditCardInfo from '../components/payments/CreditCardInfo';
 
 const PaymentPage = () => (
   <Container>
@@ -17,10 +17,10 @@ const PaymentPage = () => (
       <CouncilSearch />
     </Row>
     <Row>
-      <CheckPaymentContainer />
+      <CheckInfo />
     </Row>
     <Row>
-      <CreditCardContainer />
+      <CreditCardInfo />
     </Row>
     <Row>
       <CartTotalContainer />
