+++ app/routes.js
@@ -7,7 +7,6 @@ import PointOfSaleApp from './containers/PointOfSaleApp';
 import PaymentPage from './containers/PaymentPage';
 import InventoryPage from './containers/InventoryPage';
 import { connect } from 'react-redux';
-import CreditCardInfo from './components/CreditCardInfo';
 import * as pages from './constants/Pages';
 
 const AuthedRoute = ({ validSignOn, component: Component, ...rest }) => (
